requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-swagger-ui==3.36.0
python-escpos==3.0a8
python-dotenv==0.19.0
click==8.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyserial==3.5